# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PMC Demo API
  version: "1.0.0"
  description: |
    PMC Demo's backend is ~50 route files covering the full 13-phase build (see README.md).
    Hand-authoring every one of the ~400 endpoints here would drift out of sync immediately and
    isn't a good use of a hardening pass — instead, this document fully specifies the
    authentication flow and one representative module per architectural pattern (a plain
    workspace-scoped CRUD module, a public unauthenticated surface, and the Phase 11/12 additions),
    and states the pattern every other module follows so a reader can extrapolate correctly.

    **Every other module's routes follow this identical shape:**
    - Base path `/api/<module>` (see `backend/src/app.ts` for the exact list of mount points).
    - Every route except `/api/auth/*` and `/api/public/*` requires the `bearerAuth` security
      scheme below.
    - Every route enforces both authentication (`authenticate` middleware) and RBAC
      (`requirePermission(module, action)` middleware) — see `docs/ARCHITECTURE.md` §2 for the
      permission model and `constants/modules.ts` for the exact module/action keys.
    - List endpoints return `{ <pluralResourceName>: [...] }`; detail/create/edit endpoints return
      `{ <singularResourceName>: {...} }`; delete endpoints return `204` with no body.
    - Every detail/edit/delete route scopes its query by `workspaceId` and returns `404` (never
...
servers:
  - url: http://localhost:4000/api
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
    TokenPair:
      type: object
      properties:
        accessToken: { type: string, description: "Short-lived JWT (default 15m), carries { sub, workspaceId, membershipId }" }
        refreshToken: { type: string, description: "Long-lived opaque token (default 7d), single-use — rotated on every /auth/refresh call" }
...
  responses:
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
  /auth/signup:
    post:
      summary: Create a user, their first workspace (fully seeded), and issue tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, workspaceName]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                workspaceName: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TokenPair"
                  - type: object
                    properties:
                      user: { $ref: "#/components/schemas/User" }
                      workspace:
                        type: object
                        properties:
                          id: { type: string }
                          name: { type: string }
                          slug: { type: string }
        "409":
          description: Email already in use
  /auth/login:
    post:
      summary: Authenticate and issue tokens for the caller's earliest active workspace membership
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TokenPair"
                  - type: object
                    properties:
                      user: { $ref: "#/components/schemas/User" }
        "401":
          description: Invalid email or password
  /auth/refresh:
    post:
      summary: Rotate a refresh token for a new access/refresh pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401":
          description: Invalid, expired, or already-used (revoked) refresh token
  /auth/switch-workspace:
    post:
      summary: Issue a fresh token pair for a different workspace the caller has an active membership in (Phase 11 agency layer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId]
              properties:
                workspaceId: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "403":
          description: Not an active member of that workspace
  /auth/me:
    get:
      summary: The caller's identity, current workspace id, and resolved effective permissions
      responses:
        "200":
          description: OK
        "401": { $ref: "#/components/responses/Unauthorized" }
...